Repository methods created

See merge request SindiT15/atm-ms!19

+++ src/Main/Java/org/example/model/Account.java
@@ -28,11 +28,11 @@ public class Account {
                 '}';
     }
 
-    public long getID() {
+    public Integer getID() {
         return ID;
     }
 
-    public void setID(long ID) {
+    public void setID(Integer ID) {
         this.ID = ID;
     }
